backend/requirements.txt
fastapi
uvicorn
yt-dlp
langchain-community
langchain-core
youtube-transcript-api
pydantic>=2.7.1
langchain-openai
langsmith
python-dotenv
langchain-ollama